Added Configuration

+++ src/main/java/com/assignment/mindbowser/repo/IEmployeeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.assignment.mindbowser.entity.Employee;
 
 @Repository
-public interface IEmployeeRepository extends JpaRepository<Employee, Integer>{
+public interface IEmployeeRepository extends JpaRepository<Employee, Long>{
 
 }
